requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
tinydb==4.8.0
litellm==1.62.1
langflow==1.2.0
pydantic==2.4.2
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
typing-extensions==4.8.0
pytz==2023.3
pytest==7.4.3
httpx==0.25.1
python-jose==3.3.0
cryptography==41.0.5
uuid==1.30 